// vendors/fmp/tools/company-info/types.d.ts
export interface ExecutiveCompensationData {
  cik: string
  symbol: string
  companyName: string
  industryTitle: string
  filingDate: string
  acceptedDate: string
  nameAndPosition: string
  year: number
  salary: number
  bonus: number
  stock_award: number
  incentive_plan_compensation: number
  all_other_compensation: number
  total: number
  url: string
}